football discussion

+++ Django/football/base/views.py
@@ -1,7 +1,12 @@
 from django.shortcuts import render, redirect
-# Create your views here.
-from .models import Room
+from django.contrib import messages
+from django.db.models import Q
+from django.contrib.auth.models import User
+from .models import Room, Topic
 from .forms import RoomForm
+from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth.decorators import login_required
+from django.http import HttpResponse
 
 
 #rooms = [
@@ -9,12 +14,61 @@ from .forms import RoomForm
    # {'id':2, 'name':'Lets design'},
    # {'id':3, 'name':'Test Test Test '},
 
-#]
+#] just for test
+
+
+
+def loginPage(request):
+
+    page = 'login'
+
+    if request.user.is_authenticated:
+        return redirect('home')
+
+    if request.method == 'POST':
+        username = request.POST.get('username')
+        password = request.POST.get('password')
+
+        #try:
+            #user = User.objects.get(username = username)
+        #except:
+           # messages.error(request, 'User does not exists')
+
+        user = authenticate(request, username=username, password = password)
+        if user is not None:
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, 'Username/Password does not exist ')
+
+    context = {'page' : page}
+    return render(request, 'base/login_register.html', context)
+
+
+
+def logoutUser(request):
+    logout(request)
+    return redirect('home')
+
+
+def registerPage(request):
+    page = 'register'
+
+    return render(request, 'base/login_register.html')
+
 
 
 def home(request):
-    rooms =  Room.objects.all()
-    context = {'rooms':rooms}
+    q = request.GET.get('q') if request.GET.get('q') != None else ''
+    rooms =  Room.objects.filter(
+        Q(topic__name__icontains = q ) |
+        Q(name__icontains = q) |
+        Q(description__icontains = q)
+    )
+    ##these two lines make sure it searches for what search query contains
+    topics = Topic.objects.all()
+    room_count = rooms.count()
+    context = {'rooms':rooms, 'topics': topics,'room_count': room_count}
     return render(request, 'base/home.html', context)
 
 
@@ -23,6 +77,7 @@ def room(request, pk):
     context = {'room': room}
     return render(request, 'base/room.html', context)
 
+@login_required(login_url='login') ###to restrict pages
 def createRoom(request):
     form = RoomForm()
     if request.method == 'POST':
@@ -33,10 +88,13 @@ def createRoom(request):
     context = {'form': form}
     return render(request, 'base/room_form.html', context)
 
-
+@login_required(login_url='login')
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room) #to find which room to edit
+
+    if request.user != room.host :
+        return HttpResponse('You are not allowed here!!!')
 
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room) ## to tell which room to edit
@@ -47,9 +105,13 @@ def updateRoom(request, pk):
     context = {'form': form}
     return render(request, 'base/room_form.html', context)
 
-
+@login_required(login_url='login')
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse('You are not allowed here!!!')
+
     if request.method == 'POST':
         room.delete()
         return redirect('home')
